% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_umlauts_to_tex.R
\name{convert_umlauts_to_tex}
\alias{convert_umlauts_to_tex}
\title{Tex Codes for German Umlauts}
\usage{
convert_umlauts_to_tex(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A string with the umlauts converted to plain TeX.
}
\description{
Convert German umlauts in a string to their plain TeX representation.
}
\examples{
string <- paste("this is \u00e4 string")
print(string)
print(convert_umlauts_to_tex(string))
}
\seealso{
Other German umlaut converters: 
\code{\link{convert_umlauts_to_ascii}()},
\code{\link{get_german_umlauts}()}
}
\concept{German umlaut converters}
