% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{get_rscript_script_path}
\alias{get_rscript_script_path}
\title{Get the Path of the '\R' Code File in Case of an '\command{Rscript}' Run}
\usage{
get_rscript_script_path()
}
\value{
A vector of \code{\link{mode}} character giving the name of the \R
code file. Will be character(0) if not in an \command{Rscript} run.
}
\description{
Retrieve the path from parsing the command line arguments of a
\command{Rscript} run.
}
\examples{
get_rscript_script_path()
}
\seealso{
Other script path getter functions: 
\code{\link{get_r_cmd_batch_script_path}()},
\code{\link{get_script_name}()},
\code{\link{get_script_path}()}
}
\concept{script path getter functions}
