% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{get_script_path}
\alias{get_script_path}
\title{Get the Path of the '\R' Code File}
\usage{
get_script_path()
}
\value{
A vector of \code{\link{length}} 1 and \code{\link{mode}}
character giving the name of the \R code file if \R was run via
\command{R CMD BATCH} or
\command{Rscript}.
}
\description{
This is just a wrapper for \code{\link{get_rscript_script_path}} and
\code{\link{get_r_cmd_batch_script_path}}.
}
\examples{
get_script_path()
}
\seealso{
Other script path getter functions: 
\code{\link{get_r_cmd_batch_script_path}()},
\code{\link{get_rscript_script_path}()},
\code{\link{get_script_name}()}
}
\concept{script path getter functions}
