% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_string.R
\name{get_session_string}
\alias{get_session_string}
\title{Get a Session String}
\usage{
get_session_string()
}
\value{
An excerpt of \code{\link{sessionInfo}} as a string.
}
\description{
I sometimes wan't to document the \R session used in a string, so a need an
excerpt of \code{\link{sessionInfo}} an \code{\link{Sys.time}}.
}
\examples{
get_session_string()
}
\seealso{
Other version functions: 
\code{\link{get_package_version}()},
\code{\link{is_r_package_installed}()},
\code{\link{is_version_sufficient}()}
}
\concept{version functions}
