% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_string.R
\name{get_unique_string}
\alias{get_unique_string}
\title{Create a Fairly Unique String}
\usage{
get_unique_string()
}
\value{
A fairly unique string.
}
\description{
I sometimes need a fairly unique string, mostly for file names, that should
start with the current date.
}
\examples{
replicate(20, get_unique_string())
}
\seealso{
Other file utilities: 
\code{\link{clipboard_path}()},
\code{\link{delete_trailing_blank_lines}()},
\code{\link{delete_trailing_whitespace}()},
\code{\link{develop_test}()},
\code{\link{file_copy}()},
\code{\link{file_modified_last}()},
\code{\link{file_save}()},
\code{\link{find_files}()},
\code{\link{get_lines_between_tags}()},
\code{\link{get_mtime}()},
\code{\link{grep_file}()},
\code{\link{is_files_current}()},
\code{\link{is_path}()},
\code{\link{paths}},
\code{\link{runsed}()},
\code{\link{search_files}()},
\code{\link{split_code_file}()},
\code{\link{touch}()}
}
\concept{file utilities}
