% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_ratio.R
\name{golden_ratio}
\alias{golden_ratio}
\title{Calculate the Golden Ratio}
\usage{
golden_ratio(x)
}
\arguments{
\item{x}{The sum of the two quantities to be in the golden ratio.}
}
\value{
A numeric vector of length 2, containing the two quantities \emph{a}
and \emph{b}, \emph{a}  being the larger.
}
\description{
Divide a length using the golden ratio.
}
\examples{
golden_ratio(10)
}
\seealso{
Other bits and pieces: 
\code{\link{char2factor}()},
\code{\link{is_difftime_less}()},
\code{\link{is_valid_primary_key}()},
\code{\link{pause}()},
\code{\link{r_cmd_install}()},
\code{\link{rownames2col}()},
\code{\link{str2num}()},
\code{\link{string2words}()},
\code{\link{strip_off_attributes}()},
\code{\link{tapply}()},
\code{\link{throw}()}
}
\concept{bits and pieces}
