% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean.R
\name{is_false}
\alias{is_false}
\title{Provide \code{isFALSE} for '\R' < 3.5.0}
\usage{
is_false(x)
}
\arguments{
\item{x}{The object to be tested.}
}
\value{
\code{\link[base:TRUE]{TRUE}} if the object is set to
\code{\link[base:FALSE]{FALSE}},
\code{\link[base:FALSE]{FALSE}}
otherwise.
}
\description{
I still use \R 3.3.3 for testing, \code{isFALSE()} was introduced in \R
3.5.0.
}
\examples{
is_false("not false")
is_false(FALSE)
}
\seealso{
Other logical helpers: 
\code{\link{get_run_r_tests}()},
\code{\link{is_batch}()},
\code{\link{is_cran}()},
\code{\link{is_force}()},
\code{\link{is_installed}()},
\code{\link{is_not_false}()},
\code{\link{is_null_or_true}()},
\code{\link{is_of_length_zero}()},
\code{\link{is_r_cmd_check}()},
\code{\link{is_r_package_installed}()},
\code{\link{is_running_on_fvafrcu_machines}()},
\code{\link{is_running_on_gitlab_com}()},
\code{\link{is_success}()},
\code{\link{is_version_sufficient}()},
\code{\link{is_windows}()}
}
\concept{logical helpers}
