% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_path.R
\name{is_path}
\alias{is_path}
\title{Check Whether an Object Contains a Valid File System Path}
\usage{
is_path(x)
}
\arguments{
\item{x}{The object.}
}
\value{
\code{\link[base:TRUE]{TRUE}} on success, 
\code{\link[base:FALSE]{FALSE}} otherwise.
}
\description{
Check Whether an Object Contains a Valid File System Path
}
\examples{
is_path(tempdir())
path <- tempfile()
is_path(path)
touch(path)
is_path(path)
}
\seealso{
Other file utilities: 
\code{\link{clipboard_path}()},
\code{\link{delete_trailing_blank_lines}()},
\code{\link{delete_trailing_whitespace}()},
\code{\link{develop_test}()},
\code{\link{file_copy}()},
\code{\link{file_modified_last}()},
\code{\link{file_save}()},
\code{\link{find_files}()},
\code{\link{get_lines_between_tags}()},
\code{\link{get_mtime}()},
\code{\link{get_unique_string}()},
\code{\link{grep_file}()},
\code{\link{is_files_current}()},
\code{\link{paths}},
\code{\link{runsed}()},
\code{\link{search_files}()},
\code{\link{split_code_file}()},
\code{\link{touch}()}
}
\concept{file utilities}
