% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_doc.R
\name{missing_docs}
\alias{missing_docs}
\alias{find_missing_see_also}
\alias{find_missing_family}
\title{Find Missing Documentation}
\usage{
find_missing_see_also(path, list_families = TRUE)

find_missing_family(path, list_families = TRUE, clean = TRUE)
}
\arguments{
\item{path}{Path to a (package) directory.}

\item{list_families}{List the function families defined so far.}

\item{clean}{Remove temporary directory?}
}
\value{
For `find_missing_see_also`: a character vector of man pages with
missing `See also` sections.

For `find_missing_family`: a character vector of function names with
missing `@family` tags.
}
\description{
For \pkg{fritools}, we make exhaustive use of categorizing functions into
families with the `See also` section of the man pages (which are generated by
the @family tags in the code files).
}
\seealso{
Other searching functions: 
\code{\link{compare_vectors}()},
\code{\link{file_modified_last}()},
\code{\link{find_files}()},
\code{\link{fromto}()},
\code{\link{grep_file}()},
\code{\link{runsed}()},
\code{\link{search_files}()},
\code{\link{search_rows}()},
\code{\link{summary.filesearch}()}
}
\concept{searching functions}
