% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{r_cmd_install}
\alias{r_cmd_install}
\title{Install a Tarball or a Directory}
\usage{
r_cmd_install(
  ...,
  path = ".",
  r = get_r_binary(),
  r_args = "--vanilla",
  try_tarball = TRUE
)
}
\arguments{
\item{...}{Arguments passed to \command{CMD INSTALL}.}

\item{path}{A path to a directory, but see \emph{try_tarball}.}

\item{r}{An R binary.}

\item{r_args}{Arguments passed to \command{R}.}

\item{try_tarball}{If \code{\link{TRUE}}, the \code{path} may be a tarball,
else it is interpreted as a path to a directory.}
}
\value{
The return value of \code{\link{system2}}.
}
\description{
\code{\link[devtools:install]{devtools::install}} by defaults first builds
the tarball. Then it calls
\code{\link[callr:rcmd]{callr::rcmd}}, which allows for a lot of options.
}
\seealso{
Other bits and pieces: 
\code{\link{char2factor}()},
\code{\link{golden_ratio}()},
\code{\link{is_difftime_less}()},
\code{\link{is_valid_primary_key}()},
\code{\link{pause}()},
\code{\link{rownames2col}()},
\code{\link{str2num}()},
\code{\link{string2words}()},
\code{\link{strip_off_attributes}()},
\code{\link{tapply}()},
\code{\link{throw}()}
}
\concept{bits and pieces}
\keyword{internal}
