% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_package.R
\name{rename_package}
\alias{rename_package}
\title{Change This Package's Name}
\usage{
rename_package(path, to)
}
\arguments{
\item{path}{The new package name.}
}
\value{
Run for its side effect.
}
\description{
Does not change the path, we just rename the package, not the repo
and we must not change git remotes!
}
\seealso{
Other package functions: 
\code{\link{get_package_version}()},
\code{\link{is_r_package_installed}()},
\code{\link{is_version_sufficient}()},
\code{\link{load_internal_functions}()}
}
\concept{package functions}
\keyword{internal}
