% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames2col.R
\name{rownames2col}
\alias{rownames2col}
\title{Add a Column Containing the Row Names to a Data Frame}
\usage{
rownames2col(x, column_name)
}
\arguments{
\item{x}{A \code{\link{data.frame}}.}

\item{column_name}{The name of the new column containing the
\code{\link{row.names}}.}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Add a Column Containing the Row Names to a Data Frame
}
\examples{
rownames2col(mtcars, column_name = "model")
}
\seealso{
Other bits and pieces: 
\code{\link{char2factor}()},
\code{\link{golden_ratio}()},
\code{\link{is_difftime_less}()},
\code{\link{is_valid_primary_key}()},
\code{\link{pause}()},
\code{\link{r_cmd_install}()},
\code{\link{str2num}()},
\code{\link{string2words}()},
\code{\link{strip_off_attributes}()},
\code{\link{tapply}()},
\code{\link{throw}()}
}
\concept{bits and pieces}
