% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{set_run_r_tests}
\alias{set_run_r_tests}
\title{Set the System Variable RUN_R_TESTS}
\usage{
set_run_r_tests(x, force = FALSE)
}
\arguments{
\item{x}{A logical, typically some function output.}

\item{force}{Overwrite the variable if already set?}
}
\value{
The value RUN_R_TESTS is set to, \code{\link{NULL}} if nothing is
done.
}
\description{
A convenience wrapper to  \code{\link{Sys.getenv}} for setting
\kbd{RUN_R_TESTS}.
}
\examples{
set_run_r_tests(is_running_on_fvafrcu_machines())
get_run_r_tests()
set_run_r_tests(TRUE, force = TRUE)
get_run_r_tests()
}
\seealso{
Other test helpers: 
\code{\link{develop_test}()},
\code{\link{get_boolean_envvar}()},
\code{\link{get_run_r_tests}()},
\code{\link{is_cran}()},
\code{\link{is_r_cmd_check}()},
\code{\link{is_running_on_fvafrcu_machines}()},
\code{\link{is_running_on_gitlab_com}()},
\code{\link{run_r_tests_for_known_hosts}()}
}
\concept{test helpers}
