% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string2words.R
\name{string2words}
\alias{string2words}
\title{Convert a Character Vector Into an Enumeration}
\usage{
string2words(x, separator = ",", last = "and", add_whitespace = TRUE)
}
\arguments{
\item{x}{A \code{\link{character}} vector.}

\item{separator}{A separator used for the enumeration.}

\item{last}{The separator used last for the enumeration.}

\item{add_whitespace}{Add whitespace after separators?}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Convert a Character Vector Into an Enumeration
}
\examples{
string2words(c("beech", "oak", "ash"))
}
\seealso{
Other bits and pieces: 
\code{\link{char2factor}()},
\code{\link{golden_ratio}()},
\code{\link{is_difftime_less}()},
\code{\link{is_valid_primary_key}()},
\code{\link{pause}()},
\code{\link{r_cmd_install}()},
\code{\link{rownames2col}()},
\code{\link{str2num}()},
\code{\link{strip_off_attributes}()},
\code{\link{tapply}()},
\code{\link{throw}()}
}
\concept{bits and pieces}
