% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vim.R
\name{vim}
\alias{vim}
\title{Edit a File With '\command{VIM}' if Possible}
\usage{
vim(...)
}
\arguments{
\item{...}{See \code{\link{file.edit}}.}
}
\value{
See \code{\link{file.edit}}.
}
\description{
Just a wrapper to \code{\link{file.edit}}, trying to use [g]vim as editor, if
installed.
}
\examples{
if (interactive()) {
    path <- file.path(tempdir(), "foo.txt")
    writeLines(c("abc", "xyz"), con = path)
    vim(path)
}
}
\seealso{
Other operating system functions: 
\code{\link{clipboard_path}()},
\code{\link{file_copy}()},
\code{\link{file_save}()},
\code{\link{get_boolean_envvar}()},
\code{\link{get_run_r_tests}()},
\code{\link{is_installed}()},
\code{\link{is_r_package_installed}()},
\code{\link{is_success}()},
\code{\link{is_windows}()},
\code{\link{view}()},
\code{\link{wipe_tempdir}()},
\code{\link{with_dir}()}
}
\concept{operating system functions}
