% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wipe_tempdir.R
\name{wipe_tempdir}
\alias{wipe_tempdir}
\title{Wipe Clean the \code{tempdir()}}
\usage{
wipe_tempdir(recreate = FALSE)
}
\arguments{
\item{recreate}{Use the method described in the examples section of
\code{\link{tempdir}} (using \code{tempdir(check = TRUE)}, this results in a
new path.)}
}
\value{
The path to the temporary directory.
}
\description{
I often need a clean temporary directory.
}
\seealso{
Other R memory functions: 
\code{\link{memory_hogs}()},
\code{\link{wipe_clean}()}

Other operating system functions: 
\code{\link{clipboard_path}()},
\code{\link{file_copy}()},
\code{\link{file_save}()},
\code{\link{get_boolean_envvar}()},
\code{\link{get_run_r_tests}()},
\code{\link{is_installed}()},
\code{\link{is_r_package_installed}()},
\code{\link{is_success}()},
\code{\link{is_windows}()},
\code{\link{view}()},
\code{\link{vim}()},
\code{\link{with_dir}()}
}
\concept{R memory functions}
\concept{operating system functions}
